+++ paper/paper.bib
@@ -5,6 +5,7 @@
   volume={77},
   number={21},
   pages={e75--e78},
+  doi = {10.1158/0008-5472.CAN-17-0629},
   year={2017},
   publisher={AACR}
 }
@@ -26,6 +27,7 @@
   volume={6},
   number={1},
   pages={8390},
+  doi={10.1038/ncomms9390},
   year={2015},
   publisher={Nature Publishing Group UK London}
 }
@@ -36,6 +38,7 @@
   author={Lin, Jia-Ren and Izar, Benjamin and Wang, Shu and Yapp, Clarence and Mei, Shaolin and Shah, Parin M and Santagata, Sandro and Sorger, Peter K},
   journal={elife},
   volume={7},
+  doi={10.7554/eLife.31657},
   year={2018},
   publisher={eLife Sciences Publications, Ltd}
 }
@@ -48,6 +51,7 @@
   volume={110},
   number={29},
   pages={11982--11987},
+  doi = {10.1073/pnas.1300136110},
   year={2013},
   publisher={National Acad Sciences}
 }
@@ -57,6 +61,7 @@
   author={Schmidt, Uwe and Weigert, Martin and Broaddus, Coleman and Myers, Gene},
   booktitle={Medical Image Computing and Computer Assisted Intervention--MICCAI 2018: 21st International Conference, Granada, Spain, September 16-20, 2018, Proceedings, Part II 11},
   pages={265--273},
+  doi={10.1007/978-3-030-00934-2_30},
   year={2018},
   organization={Springer}
 }
@@ -70,7 +75,7 @@
   volume={18},
   number={3},
   pages={259--278},
-  doi = {https://doi.org/10.2307/2412323},
+  doi = {10.2307/2412323},
   year={1969},
   publisher={Society of Systematic Zoology}
 }
@@ -83,6 +88,7 @@
   volume={174},
   number={4},
   pages={968--981},
+  doi={10.1016/j.cell.2018.07.010},
   year={2018},
   publisher={Elsevier}
 }
@@ -94,6 +100,7 @@
   volume={20},
   number={4},
   pages={436--442},
+  doi = {10.1038/nm.3488},
   year={2014},
   publisher={Nature Publishing Group US New York}
 }
