+++ js/histojs/designFunctionsv4.js
@@ -44,20 +44,26 @@
              {id: "2", value: "Styx", hostAPI: "https://styx.neurology.emory.edu/girder/api/v1/"}]
    *
    * findObjectByKeyValue( dict, 'id', "2");
-   * // => Object { id: "2", value: "Styx", hostAPI: "https://styx.neurology.emory.edu/girder/api/v1/" }
-   *
+   *  => Object { id: "2", value: "Styx", hostAPI: "https://styx.neurology.emory.edu/girder/api/v1/" }
    *
    * findObjectByKeyValue( dict, 'id', "2", 'INDEX');
    *  => 1
-   *
+   * 
+   * await findObjectByKeyValueTFJS( dict, 'id', "2")
+   * => Object { id: "2", value: "Styx", hostAPI: "https://styx.neurology.emory.edu/girder/api/v1/" }
    */ 
 
-    findObjectByKeyValue = (array, key, value, returnType='DATA') => {   
-        const obj = array.filter( entry => entry[key] === value)[0];
-        return obj ? ( returnType === 'DATA' ? obj : ( returnType === 'INDEX' ? array.indexOf(obj) : null )) : null;
+    findObjectByKeyValue = (arr, key, value, returnType = 'DATA') => {   
+        const obj = arr.filter( entry => entry[key] === value)[0];
+        return obj ? ( returnType === 'DATA' ? obj : ( returnType === 'INDEX' ? arr.indexOf(obj) : null )) : null;
     }    
 
-     
+    findObjectByKeyValueTFJS = async (arr, key, value) => {   
+        const result = await tf.data.array(arr).filter( entry => entry[key] === value).toArray();
+        return (result.length > 1) ? result : result.length ? result[0] : null;
+    }    
+
+ 
   /**
    * Remove element or record from array
    *
@@ -74,6 +80,9 @@
    * removeArrayElem(['a', 'b', 'c', 'd'], 'c')
    *  => ['a', 'b', 'd']
    *
+   *  await removeArrayElemTFJS([1, 2, 3, 4], 3)
+   *  => [1, 2, 4]
+   *  NOTE: TFJS version for numbers only
    */    
 
     removeArrayElem = (array, element) => { 
@@ -82,6 +91,17 @@
             array.splice(index, 1);
         }
     }
+
+    removeArrayElemTFJS = async (array, element) => {
+        const tensor = tf.tensor(array);
+        const mask = tf.notEqual(tensor, element);
+        const result = await tf.booleanMaskAsync(tensor, mask); 
+        return result.arraySync();
+    }
+
+    removeArrayElem_v2 = (array, element) => { 
+        return array.filter(elm => elm !== element);
+    }  
 
 
  /**
@@ -143,11 +163,22 @@
    * @returns {Array} arrOfObj2 - The second array of objects to be merged.
    * @example
    *
-   * mergeArrayOfObjByKey( "id", [ {id:"spx-1", Type:"Tumor"}, 
-   *                               {id:"spx-7", Type:"Immune"} ], 
+   *   mergeArrayOfObjByKey( "id", [ {id:"spx-1", Type:"Tumor" }, 
+   *                                 {id:"spx-7", Type:"Immune"} ], 
    *                                                            [{id : "spx-1", area: 250,  solidity: 0.95}, 
    *                                                             {id : "spx-3", area: 150,  solidity: 0.85}, 
    *                                                             {id : "spx-7", area: 100,  solidity: 0.80} ])
+   *
+   * => [{ id: "spx-1", Type:"Tumor", area: 250, solidity: 0.95 }, 
+   *     { id: "spx-7", Type: "immune", area: 100, solidity: 0.80 }]
+   *
[…]
