+++ js/histojs/designFunctionsv4.js
@@ -72,35 +72,47 @@
    * @since 1.0.0
    * @version 1.0.0
    * @category Array
-   * @param {Array} array The array to process.
+   * @param {Array} arr -  The array to process.
    * @param {*} element The value to inspect and remove.
    * @returns {Array} 
    * @example
    *
    * removeArrayElem(['a', 'b', 'c', 'd'], 'c')
    *  => ['a', 'b', 'd']
-   *
-   *  await removeArrayElemTFJS([1, 2, 3, 4], 3)
+   *---------------------------------------------*
+   *  await removeArrayElemTFJS( [1, 2, 3, 4], 3)
    *  => [1, 2, 4]
-   *  NOTE: TFJS version for numbers only
+   *
+   *  await removeArrayElemTFJS(['a', 'b', 'c', 'd'], 'c')
+   *  => ['a', 'b', 'd']
+   *---------------------------------------------*
+   *  await removeArrayElemTFJSV2( [1, 2, 3, 4], 3)
+   *  => [1, 2, 4]
+   *  NOTE: removeArrayElemTFJSV2 for numbers only
+   *
    */    
 
-    removeArrayElem = (array, element) => { 
-        const index = array.indexOf(element);
+    removeArrayElem = (arr, element) => { 
+        const index = arr.indexOf(element);
         if(index >= 0 ) {
-            array.splice(index, 1);
+            arr.splice(index, 1);
         }
     }
 
-    removeArrayElemTFJS = async (array, element) => {
-        const tensor = tf.tensor(array);
+    removeArrayElemTFJS = async(arr, element) => { 
+        return await tf.data.array(arr).filter(elm => elm !== element).toArray();
+    }      
+
+    
+    removeArrayElemTFJSV2 = async (arr, element) => {
+        const tensor = tf.tensor(arr);
         const mask = tf.notEqual(tensor, element);
         const result = await tf.booleanMaskAsync(tensor, mask); 
         return result.arraySync();
     }
 
-    removeArrayElem_v2 = (array, element) => { 
-        return array.filter(elm => elm !== element);
+    removeArrayElem_v2 = (arr, element) => { 
+        return arr.filter(elm => elm !== element);
     }  
 
 
@@ -258,6 +270,13 @@
    *                               {id:"spx-7", Type:"Immune"} ])
    *
    *  => Object { "spx-1": {id: "spx-1", Type: "Tumor"}, "spx-7": {id: "spx-7", Type: "Immune"} }
+   *
+   *
+   * await array2ObjWithHashKeyTFJS( "id", [ {id:"spx-1", Type:"Tumor"}, 
+   *                                         {id:"spx-7", Type:"Immune"} ])
+   *
+   *  => Object { "spx-1": {id: "spx-1", Type: "Tumor"}, "spx-7": {id: "spx-7", Type: "Immune"} }
+   *
    */ 
 
     array2ObjWithHashKey = (hashKey, arrayOfObjs) => {
@@ -269,12 +288,31 @@
                  convertedObject[ object[hashKey] ] = object; 
             });  
 
-            return   convertedObject;  
+            return convertedObject;  
+
         } else {
             triggerHint("Not a valid array of objects .. ") 
             return null; 
         }          
     }
+
+    array2ObjWithHashKeyTFJS = async(hashKey, arrayOfObjs) => {
+        if(Object.keys(arrayOfObjs).length) {
+            // create hash folarger arrayr the 
+            let convertedObject = {};
+
+            await tf.data.array(arrayOfObjs).forEachAsync(object => { 
[…]
