+++ js/histojs/designFunctionsv4.js
@@ -40,28 +40,39 @@
    * @returns {object | number} 
    * @example
    *
-   * dict = [{id: "1", value: "Girder", hostAPI: "http://dermannotator.org:8080/api/v1" },
-             {id: "2", value: "Styx", hostAPI: "https://styx.neurology.emory.edu/girder/api/v1/"}]
+   * dict = [ {id: "1", value: "Girder"}, {id: "2", value: "Styx"} ]
    *
    * findObjectByKeyValue( dict, 'id', "2");
-   *  => Object { id: "2", value: "Styx", hostAPI: "https://styx.neurology.emory.edu/girder/api/v1/" }
+   *  => Object { id: "2", value: "Styx" }
    *
    * findObjectByKeyValue( dict, 'id', "2", 'INDEX');
    *  => 1
-   * 
+   *
+   * findObjectByKeyValue( allTilesFeaturesAndClassification, 'id', "spx-105401");
+   *
+   * => Object { label: 105401, features: (5) […], area: 152, eccentricity: 0.28, orientation: 138, perimeter: 46.74,
+   *          extent: 0.63, solidity: 0.96, major_axis_length: 14.37, minor_axis_length: 13.78, … }
+   *---------------------------------------------------------------*
    * await findObjectByKeyValueTFJS( dict, 'id', "2")
-   * => Object { id: "2", value: "Styx", hostAPI: "https://styx.neurology.emory.edu/girder/api/v1/" }
+   * => Object { id: "2", value: "Styx" }
+   *
+   * await findObjectByKeyValueTFJS( allTilesFeaturesAndClassification, 'id', "spx-105401")
+   * => Object { label: 105401, features: (5) […], area: 152, eccentricity: 0.28, orientation: 138, perimeter: 46.74, 
+   *             extent: 0.63, solidity: 0.96, major_axis_length: 14.37, minor_axis_length: 13.78, … }
+   *
+   * NOTE: findObjectByKeyValueTFJS slower than findObjectByKeyValue.
    */ 
 
-    findObjectByKeyValue = (arr, key, value, returnType = 'DATA') => {   
+    findObjectByKeyValue = (arr, key, value, returnType = 'DATA') => { //fast   
         const obj = arr.filter( entry => entry[key] === value)[0];
         return obj ? ( returnType === 'DATA' ? obj : ( returnType === 'INDEX' ? arr.indexOf(obj) : null )) : null;
     }    
 
-    findObjectByKeyValueTFJS = async (arr, key, value) => {   
+
+    findObjectByKeyValueTFJS = async (arr, key, value) => { //slow
         const result = await tf.data.array(arr).filter( entry => entry[key] === value).toArray();
         return (result.length > 1) ? result : result.length ? result[0] : null;
-    }    
+    }   
 
  
   /**
@@ -77,18 +88,24 @@
    * @returns {Array} 
    * @example
    *
+   * arrOfObjs = [{ Type: "1", val: 1}, { Type: "2", val: 2}]
+   *
    * removeArrayElem(['a', 'b', 'c', 'd'], 'c')
    *  => ['a', 'b', 'd']
+   *
+   * removeArrayElem(arrOfObjs, arrOfObjs[1])
+   *
+   * => [{ Type: "1", val: 1}]
    *---------------------------------------------*
    *  await removeArrayElemTFJS( [1, 2, 3, 4], 3)
    *  => [1, 2, 4]
    *
    *  await removeArrayElemTFJS(['a', 'b', 'c', 'd'], 'c')
    *  => ['a', 'b', 'd']
-   *---------------------------------------------*
-   *  await removeArrayElemTFJSV2( [1, 2, 3, 4], 3)
-   *  => [1, 2, 4]
[…]
