+++ README.md
@@ -7,7 +7,7 @@
  [![MIT-License ](https://img.shields.io/badge/license-MIT-green)](https://github.com/Mmasoud1/HistoJS/blob/main/LICENSE) [![Python ](https://img.shields.io/badge/python-3.5%20%7C%203.6%20%7C%203.7-blue)]()[![JS ](https://img.shields.io/badge/Types-JavaScript-blue)]()
 [![HW ](https://img.shields.io/badge/HardWare-GPU-green)]()
 
-**[Demo](https://mmasoud1.github.io/HistoJS) &emsp;  [Updates](#Updates) &emsp; [Doc](https://github.com/mmasoud1/HistoJS/wiki/)**
+**[Demo](https://mmasoud1.github.io/HistoJS) &emsp;  [Updates](#Updates) &emsp; [Doc](https://github.com/mmasoud1/HistoJS/wiki/) &emsp; [News!](#News)**
 
 </div>
 
@@ -48,6 +48,17 @@ https://github.com/Mmasoud1/HistoJS/assets/36082112/7ba5c733-3b31-47e9-9cc4-e9fb
 
 <br>
 
+## News
+
+
+* HistoJS abstract is accepted for poster presentation in the [OHBM](https://www.humanbrainmapping.org/) 2024 Annual Meeting in Seoul, Korea.
+
+<div align="center">
+   <img src="https://github.com/neuroneural/brainchop/blob/master/Demo/News/OHBM_2024.png"  width="40%">
+</div>
+
+<br>
+<br>
 
 <div align="center">
 
