### 2024-02-01 #6217 @mooniean
Hi, @cinnetcrash @alexriss @suresh-lab - would you be available and interested in reviewing this submission for the Journal of Open Source Software?

>  HistoJS: Web-Based Analytical Tool for Advancing Multiplexed Images #6217

Thanks in advance!
