### 2024-02-03 #6317 @editorialbot
[REVIEW]: HistoJS: Web-Based Analytical Tool for Advancing Multiplexed Images
**Submitting author:** @Mmasoud1 (<a href="http://orcid.org/0000-0002-5365-242X">Mohamed Masoud</a>)
**Repository:** https://github.com/Mmasoud1/HistoJS
**Branch with paper.md** (empty if default branch):
**Version:** v1.1.0
