### 2024-02-04 #6317 @alexriss
It seems fine from my side - the usual conda/pip challenges during installation, but it worked in the end.
