### 2024-02-05 #6317 @mooniean
Brilliant! Thank you for your time and expertise, reviewers!

Before the next step, @Mmasoud1, there's some minor format adjustments on the paper:
- there's a reference to [3-4] in the Summary which I'm not sure what it refers to
- Typo in line 28: "remain" -> "remains"
- Missing word in line 91: "it can also [be] deployed"
- In the code availability, there's some markdown issues with the link to the GitHub

After this, we'll go to the post-review checklist!
