+++ paper/paper.bib
@@ -151,7 +151,17 @@
 }
 
 
-
+@article{credit,
+  author  = {Brand, Amy and Allen, Liz and Altman, Micah and Hlava, Marjorie and Scott, Jo},
+  title   = {Beyond authorship: attribution, contribution, collaboration, and credit},
+  journal = {Learned Publishing},
+  volume  = {28},
+  number  = {2},
+  pages   = {151-155},
+  doi     = {10.1087/20150211},
+  url     = {https://onlinelibrary.wiley.com/doi/abs/10.1087/20150211},
+  year    = {2015}
+}
 
 
 
