+++ paper/paper.md
@@ -29,14 +29,14 @@ bibliography: paper.bib
 
 # Summary
 
-Advances in multiplexed imaging technologies enable us to capture single-cell proteomics and transcriptomics data in unprecedented detail and with high spatial resolution at the single-cell level. This large volume of image data presents a challenge in accurately isolating and quantifying distinct cell types to understand disease complexity, neurological disorders, potential biomarkers, and druggable targets for drug development. Therefore, there is an incremental demand and challenges to developing and validating cutting-edge quantitative image analysis tools for diagnosis, prognosis, and therapy response prediction and assessment in neurological and oncological diseases. [HistoJS](https://github.com/Mmasoud1/HistoJS) is a newly developed web-based tool that aims to overcome the challenges of utilizing highly-multiplexed immunofluorescence (HMIF) images [3-4] for Spatial biology research. It provides open-source and extensible tools for analyzing spatial-molecular patterns, enabling a deeper view of the single-cell spatial relationship, along with machine learning algorithms in an easy-to-use interactive interface for the Biomedical community. 
+Advances in multiplexed imaging technologies enable us to capture spatial single-cell proteomics and transcriptomics data in unprecedented detail and with high spatial resolution. This large volume of image data presents a challenge in accurately isolating and quantifying distinct cell types to understand disease complexity, neurological disorders, potential biomarkers, and targets for drug development. Therefore, there is an incremental demand and challenges to developing and validating cutting-edge quantitative image analysis tools for diagnosis, prognosis, and therapy response prediction and assessment in neurological and oncological diseases. [HistoJS](https://github.com/Mmasoud1/HistoJS) is a newly developed web-based tool that aims to overcome the challenges of utilizing highly-multiplexed immunofluorescence (mIF) images [lin2015highly] for spatial biology research. It provides open-source and extensible tools for analyzing spatial-molecular patterns, enabling a deeper view of the single-cell spatial relationship, along with machine learning algorithms in an easy-to-use interactive interface for the biomedical community. 
 
 
 
 # Statement of need
 
-Single-cell data has the potential to enhance the understanding of biological systems and enable the understanding of disease mechanisms and reactions. The data resulting from single-cell sequencing and multiplexed imaging technologies has a distinctive capability and can be used for cell phenotyping, clustering, and understanding its landscape.
[…]
