### 2024-02-06 #6317 @Mmasoud1
Thank you @mooniean, I applied all suggested fixes in the article and reviewed all the text again, adding one more reference for the CRediT taxonomy and Author contributions section.
