### 2024-02-06 #6317 @Mmasoud1
Thank you @mooniean,  I finished the additional tasks above, double checked authors and affiliations (including ORCIDs),
 please kindly have the needed information:

**Release version:**       v1.1.0

**Zenodo DOI:**       10.5281/zenodo.10626533

**MIT license** with Github [repo](https://github.com/Mmasoud1/HistoJS?tab=MIT-1-ov-file#readme)  and [Zenodo](https://zenodo.org/doi/10.5281/zenodo.10626533)
