### 2024-02-07 #6317 @Mmasoud1
Thank you, @Kevin-Mattheus-Moerman; we are thrilled to have the HistoJS paper published with JOSS.  :fire: :tada:

We would like to take this opportunity to express our gratitude again to our reviewers, @alexriss and @cinnetcrash. Your valuable time and effort are highly appreciated, and we could not have achieved this without your help.

Lastly, special thanks to @mooniean for her fantastic editing. The reviewing and editing process was outstanding.  :100: 👌

:heart:
